requirements.txt
<<<<<<< Updated upstream
pygame~=2.0.0.dev8
=======
tkinter
mysql-connector-python
py_mainform_2
>>>>>>> Stashed changes
